Added a better error message for when 'bsub' is missing

+++ plumbing/cmd_test.py
@@ -7,15 +7,10 @@ import os
 
 # Internal modules #
 from plumbing import command
+from plumbing.common import check_executable
 
-# Unittesting module #
-try:
-    import unittest2 as unittest
-except ImportError:
-    import unittest
-
-# Nosetest flag #
-__test__ = True
+# Testing module #
+from nose.plugins.skip import SkipTest
 
 ###################################################################################
 @command
@@ -24,37 +19,34 @@ def touch(filename):
             "return_value": filename}
 
 ###################################################################################
-class Test(unittest.TestCase):
-    def runTest(self):
-        # Simple touch #
-        name = "myfile"
-        got = touch(name)
-        self.assertEqual(got, name)
-        self.assertTrue(os.path.exists(name))
-        os.remove(name)
-        # Parallel touch #
-        name1 = "myfile1"
-        name2 = "myfile2"
-        future1 = touch.parallel(name1)
-        future2 = touch.parallel(name2)
-        got1 = future1.wait()
-        got2 = future2.wait()
-        self.assertEqual(got1, name1)
-        self.assertEqual(got2, name2)
-        self.assertTrue(os.path.exists(name1))
-        self.assertTrue(os.path.exists(name2))
-        os.remove(name1)
-        os.remove(name2)
-        # LSF touch #
-        path = "/scratch/cluster/daily/%s/plumbing_test" % os.environ['USER']
-        directory = os.path.dirname(path)
-        if not os.path.exists(directory): os.makedirs(directory)
-        future = touch.lsf(path)
-        got = future.wait()
-        self.assertEqual(got, path)
-        self.assertTrue(os.path.exists(path))
-        os.remove(path)
+def test_simple_touch():
+    name = "myfile"
+    got = touch(name)
+    assert got == name
+    assert os.path.exists(name)
+    os.remove(name)
 
-###################################################################################
-if __name__ == '__main__':
-    Test().runTest()
+def test_parallel_touch():
+    name1 = "myfile1"
+    name2 = "myfile2"
+    future1 = touch.parallel(name1)
+    future2 = touch.parallel(name2)
+    got1 = future1.wait()
+    got2 = future2.wait()
+    assert got1 == name1
+    assert got2 == name2
+    assert os.path.exists(name1)
+    assert os.path.exists(name2)
+    os.remove(name1)
+    os.remove(name2)
+
+def test_lsf_touch():
+    if not check_executable('bsub'): raise SkipTest
+    path = "/scratch/cluster/daily/%s/plumbing_test" % os.environ['USER']
+    directory = os.path.dirname(path)
+    if not os.path.exists(directory): os.makedirs(directory)
+    future = touch.lsf(path)
+    got = future.wait()
+    assert got == path
+    assert os.path.exists(path)
+    os.remove(path)
